alternator: split export executor code
